Bulk importation of Alexey's work, part 2

+++ thebrig_install.sh
@@ -3,7 +3,8 @@
 # define our bail out shortcut function anytime there is an error - display the error message, then exit
 # returning 1.
 exerr () { echo -e "$*" >&2 ; exit 1; }
-
+startfolder=`pwd`
+echo $startfolder > /tmp/thebriginstaller
 # This first checks to see that the user has supplied an argument
 if [ ! -z $1 ]; then
     # The first argument will be the path that the user wants to be the root folder.
@@ -16,7 +17,10 @@ if [ ! -z $1 ]; then
         echo "Attempting to create a new destination directory....."
         mkdir -p $BRIG_ROOT || exerr "ERROR: Could not create directory!"
     fi
-    cd $BRIG_ROOT || exerr "ERROR: Could not access install directory!"
+	mkdir -p temporary || exerr "ERROR: Could not create install directory!"
+	echo $BRIG_ROOT > /tmp/thebrig.tmp
+	cd temporary || exerr "ERROR: Could not access install directory!"
+#    cd $BRIG_ROOT || exerr "ERROR: Could not access install directory!"
 else
 # We are here because the user did not specify an alternate location. Thus, we should use the 
 # current directory as the root.
@@ -30,51 +34,55 @@ else
 fi
 
 
-# This is the script to initially install thebrig
-# It first fetches the zip of the most recent version from github
-# and then extracts it.
-
 # Fetch the master branch as a zip file
 echo "Retrieving the most recent version of TheBrig"
-fetch https://github.com/fsbruva/thebrig/archive/master.zip || exerr "ERROR: Could not write to install directory!"
+fetch https://github.com/alexey1234/mythebrig/archive/master.zip || exerr "ERROR: Could not write to install directory!"
+
 
 # Extract the files we want, stripping the leading directory, and exclude
 # the git nonsense
 echo "Unpacking the tarball..."
 tar -xvf master.zip --exclude='.git*' --strip-components 1
-
 # Get rid of the tarball
 rm master.zip
+/usr/local/bin/php-cgi -f conf/bin/change_ver.php
 
-# Rename some files we have so there is only one bin/ftp
-if [ `uname -p` = "amd64" ]; then
-    echo "Renaming 64 bit ftp binary"
-    mv conf/bin/ftp_amd64 conf/bin/ftp
-    rm conf/bin/ftp_i386
-else
-    echo "Renaming 32 bit ftp binary"
-    mv conf/bin/ftp_i386 conf/bin/ftp
-    rm conf/bin/ftp_amd64
-fi
-
-# Place the path (of the current directory) within a file for the intial
-# run of the extension
-touch /tmp/thebrig.tmp
-echo $BRIG_ROOT > /tmp/thebrig.tmp
-
-# Copy all the requisite files to be used into the /usr/local/www folders as needed
-mkdir -p /usr/local/www/ext/thebrig
-cp $BRIG_ROOT/conf/ext/thebrig/* /usr/local/www/ext/thebrig
-cd /usr/local/www
-# For each of the php files in the extensions folder
-for file in /usr/local/www/ext/thebrig/*.php
-do
+file="/tmp/thebrigversion"
+if [ -f "$file" ]
+then
+	echo "Thebrig install/update"
+		if [ `uname -p` = "amd64" ]; then
+			echo "Renaming 64 bit ftp binary"
+			mv conf/bin/ftp_amd64 conf/bin/ftp
+			rm conf/bin/ftp_i386
+		else
+			echo "Renaming 32 bit ftp binary"
+			mv conf/bin/ftp_i386 conf/bin/ftp
+			rm conf/bin/ftp_amd64
+		fi
+	cp -r * $BRIG_ROOT/
+	mkdir -p /usr/local/www/ext/thebrig
+	cp $BRIG_ROOT/conf/ext/thebrig/* /usr/local/www/ext/thebrig
+	cd /usr/local/www
+	# For each of the php files in the extensions folder
+	for file in /usr/local/www/ext/thebrig/*.php
+	do
 	# Check if the link is alredy there
-	if [ -e "${file##*/}" ]; then
-		rm "${file##*/}"
-	fi
-	# Create link
-	ln -s "$file" "${file##*/}"
-done
-
-echo "Congratulations! Refresh to see a new tab under \" Extensions\"!"
+		if [ -e "${file##*/}" ]; then
+			rm "${file##*/}"
+		fi
+			# Create link
+		ln -s "$file" "${file##*/}"
+		done
+	echo "Congratulations! Thebrig was updated/installed . Navigate to rudimentary config and push Save "
+else
+	echo "You use fresh version"
+fi
+# Clean after work
+cd $startfolder
+rm -Rf temporary/*
+rmdir temporary
+rm /tmp/thebriginstaller
+rm /tmp/thebrigversion
+currentdate=`date -j +"%Y-%m-%d %H:%M:%S"`
+echo "[$currentdate]: TheBrig installer!: installer: install/upgrade action successfull" >> $BRIG_ROOT/thebrig.log
